Fold pr-status into the pr module as a default-on `status` option
The standalone pr-status module duplicated all of pr's branch
resolution, caching, and background-refresh machinery and spawned a
second `gh` process. Merge it into the pr module: a single
`gh pr view` call now fetches both the PR state and the check-status
rollup, cached together.

The `pr` config segment becomes an object with a `status` flag
(default on): `{"pr": {"status": false}}` shows just the PR number.
The status colours move under the `PrScheme` trait and the `pr` theme
key (status_success_fg, etc.).

Removes the PrStatus segment/module/scheme and the hidden
refresh-pr-status subcommand.
